backend/agent_env/requirements.txt
# LangChain for agentic video generation
langchain==0.1.16
langchain-nvidia-ai-endpoints==0.0.4
langchain-core==0.1.48
numpy<2

# Video generation with Manim
manim==0.17.3
moviepy==1.0.3
ffmpeg-python==0.2.0

# Other dependencies
python-dotenv==1.0.1
